Adding http namespace
