--- sdk/core/azure-core/src/az_http_request.cpp
@@ -1,31 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// SPDX-License-Identifier: MIT
-
-#pragma once
-
-#include <contract.hpp>
-
-namespace azure
-{
-namespace core
-{
-    class az_http_request
-    {
-    private:
-        /* data */
-    public:
-        az_http_request();
-        ~az_http_request();
-    };
-    
-    az_http_request::az_http_request()
-    {
-    }
-    
-    az_http_request::~az_http_request()
-    {
-    }
-    
-
-} // namespace core
-} // namespace azure
