+++ sdk/core/azure-core/src/http/az_http_request.cpp
@@ -1,0 +1,28 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// SPDX-License-Identifier: MIT
+
+#include <contract.hpp>
+
+namespace azure
+{
+namespace core
+{
+namespace http
+{
+class az_http_request
+{
+
+private:
+  /* data */
+public:
+  az_http_request();
+  ~az_http_request();
+};
+
+az_http_request::az_http_request() {}
+
+az_http_request::~az_http_request() {}
+
+} // namespace http
+} // namespace core
+} // namespace azure
